refactor CtrlClient
